requirements.txt
# TODO: Update with exact module version
numpy
torch>=1.7
opencv_python
loguru
scikit-image
tqdm
torchvision>=0.10.0
Pillow
thop
ninja
tabulate
tensorboard
lap
# motmetrics # now we use the local and customized version instead
filterpy
h5py

# verified versions
# onnx==1.8.1
# onnxruntime==1.8.0
# onnx-simplifier==0.3.5
